Updated with Embedded videos using iFrame
Embedded videos from youtube to act as placeholder for videos that will be made using Adobe Captivate. HTML and CSS are still being updated but this is a basic idea of the format we want to use. There needs to be a scroll (overflow) installed on this code. Currently, it is a very long column. It would normally have the size at about 50 - 75% max screen height and 20 - 30% max screen width.